requirements/prod.txt
pip==23.0.1
uvicorn==0.27.1
fastapi==0.110.0
alembic==1.13.1
SQLAlchemy==2.0.27
psycopg2==2.9.9
passlib==1.7.4
pytz==2024.1
python-jose==3.3.0
python-decouple==3.8
python-multipart==0.0.9
dependency-injector==4.41.0
user_identity
user_identity_cqrs_contract
framework
myeshop